pyproject.toml
[tool.poetry]
name = "djangodds"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.4"
djangorestframework = "^3.16.0"
dotenv = "^0.9.9"
pillow = "^11.3.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
psycopg = "^3.2.9"


[tool.poetry.group.lint.dependencies]
mypy = "^1.17.0"
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
